+++ src/item/item.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+import { Category } from '../entities/category.entity';
+import { Container } from '../entities/container.entity';
+import { Item } from '../entities/item.entity';
+import { ItemController } from './item.controller';
+import { ItemService } from './item.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Item, Container, Category])],
+  controllers: [ItemController],
+  providers: [ItemService],
+  exports: [ItemService],
+})
+export class ItemModule {}